who can touch _content
