+++ src/index.js
@@ -8,6 +8,7 @@ import 'prosemirror/src/collab'
 import GridSchema from './schema'
 import GridToDoc from './convert/grid-to-doc'
 import DocToGrid from './convert/doc-to-grid'
+import {isMediaType} from './convert/types'
 
 import PluginWidget from './plugins/widget.js'
 import UrlEmbedder from './plugins/url-embedder.js'
@@ -61,6 +62,21 @@ export default class Ed {
     this.pm.off('change')
     this.container.innerHTML = ''
   }
+  getBlock (id) {
+    return getItemWithId(this._content, id)
+  }
+  updateMediaBlock (block) {
+    // Widget plugin calls this to update a block in the content array
+    // Only media blocks can use this.
+    if (!block || !block.id || !block.type || !isMediaType(block.type)) {
+      throw new Error('Cant update this block')
+    }
+    let index = getIndexWithId(this._content, block.id)
+    if (index === -1) return
+
+    // MUTATION
+    this._content.splice(index, 1, block)
+  }
   set content (content) {
     // Cache the content object that we originally get from the API.
     // We'll need the content and block metadata later, in `get content`.
@@ -76,4 +92,22 @@ export default class Ed {
     let doc = this.pm.getContent()
     return DocToGrid(dom, doc, this._content)
   }
+}
+
+// Util
+
+function getIndexWithId (array, id) {
+  for (let i = 0, len = array.length; i < len; i++) {
+    let item = array[i]
+    if (item.id === id) {
+      return i
+    }
+  }
+  return -1
+}
+
+function getItemWithId (array, id) {
+  let index = getIndexWithId(array, id)
+  if (index === -1) return
+  return array[index]
 }
